requirements.txt
Pillow
pillow-heif
boto3
psycopg2-binary
python-dotenv
